+++ src/main/java/com/nexmo/client/voice/ncco/ConnectNcco.java
@@ -24,15 +24,16 @@ package com.nexmo.client.voice.ncco;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.nexmo.client.voice.PhoneEndpoint;
+import com.nexmo.client.voice.Endpoint;
 import com.nexmo.client.voice.MachineDetection;
+import com.nexmo.client.voice.PhoneEndpoint;
 
 @JsonInclude(value = JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ConnectNcco implements Ncco {
     private static final String ACTION = "connect";
 
-    private PhoneEndpoint[] endpoint;
+    private Endpoint[] endpoint;
     private String from = null;
     private Integer timeout = null;
     private Integer limit = null;
@@ -40,28 +41,28 @@ public class ConnectNcco implements Ncco {
     private String[] eventUrl = null;
     private String eventMethod = null;
 
-    public ConnectNcco(@JsonProperty("endpoint") PhoneEndpoint[] endpoint) {
+    public ConnectNcco(@JsonProperty("endpoint") Endpoint[] endpoint) {
         this.endpoint = endpoint;
     }
 
-    public ConnectNcco(PhoneEndpoint endpoint) {
-        this(new PhoneEndpoint[]{endpoint});
+    public ConnectNcco(Endpoint endpoint) {
+        this(new Endpoint[]{endpoint});
     }
 
     public ConnectNcco(String number) {
         this(new PhoneEndpoint(number));
     }
 
-    public PhoneEndpoint[] getEndpoint() {
+    public Endpoint[] getEndpoint() {
         return endpoint;
     }
 
-    public void setEndpoint(PhoneEndpoint endpoint) {
-        setEndpoint(new PhoneEndpoint[]{endpoint});
+    public void setEndpoint(Endpoint endpoint) {
+        setEndpoint(new Endpoint[]{endpoint});
     }
 
     @JsonProperty("endpoint")
-    public void setEndpoint(PhoneEndpoint[] endpoint) {
+    public void setEndpoint(Endpoint[] endpoint) {
         this.endpoint = endpoint;
     }